migrate login.py to cli.py
